Work wih neighbour search

+++ inc/partition.h
@@ -17,12 +17,12 @@ namespace sibernetic{
             size_t start;
             size_t end;
             size_t start_cell_id;
-            size_t end_cell_id;
-            size_t start_ghost_cell_id;
-            size_t end_ghost_cell_id;
-            size_t size() const { return end - start; }
-            size_t cell_count() const { return end_cell_id - start_cell_id; }
-            size_t total_cell_count() const { return end_ghost_cell_id - start_ghost_cell_id; }
+            unsigned int end_cell_id;
+            unsigned int start_ghost_cell_id;
+            unsigned int end_ghost_cell_id;
+            unsigned int size() const { return end - start; }
+            unsigned int cell_count() const { return end_cell_id - start_cell_id; }
+            unsigned int total_cell_count() const { return end_ghost_cell_id - start_ghost_cell_id; }
         };
     }
 }
